Finally support for /me.
Due to the way libotr works these are handled inline, i.e. they are not
send out as IRC CTCP action messages but as regular private messages.
However, the prefix '/me ' is prepended to them and when irssi-otr
receives a message with that prefix it will inject it as a /me locally
so that this is transparent to the user. Interoperability with other
clients is of course an issue here but for now this seems to be the best
solution.