Reengineering part of the module
This is a *big* commmit but was needed. The context data is no longer
used.

Rename a bunch of data structure to respect name space and have more
meaning to the human eye.

Fix the status bar redraw but still have a problem with the state of the
authentication. This should actually be done in the otr ops callback but
irssi is not helping much about that needing to send a signal to redraw
the status bar...

There is a bunch of other fixes but it's for the best. There is still
features to fix and test so we are still in heavy development mode.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>