Add fingerprint encrypted msgstate check function
This new function checks if one of the fingerprint's context(es) is in
the encrypted state.

Also fix a bad assignment when allocating the peer otr context.

This commit fixes an issue with the otr forget command that was using
the master context thus always having a plaintext state.

There is also a NULL check added to the otr contexts command so we don't
deref. a NULL pointer.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>