Fix: Finally fix the build system
Autoconf/make was clashing with the real libotr.so of the system so the
module object was renamed to "otr.so".

Also, this commit copies the libotr.m4 from the upstream version and add
it locally so it does not have to use the aclocal/ directory of the
system when bootstrapping.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>