Refactor OTR key generation
Use a thread for key generation. However, libperl used by irssi does not
seems to like very much emitting irssi signal inside a thread which
triggers segfaults. So, for this reason, we can not print anything
inside the thread. This is why the key gen. check is added at each event
of the modules so we can notify at some point in time the key
completion. This is quite a big limitation but for now it's the only way
I found to not trigger irssi bad behaviors in a forked process or inside
the thread.

Also, the genkey abort option is removed since when using the libgcrypt
in a multithread program, gcrypt mutexes need to be initialized then
used to lock down the key generation which basically stalls irssi until
completion. So, not having an abort command avoids the non trivial task
of handling thread cancellation and libgcrypt internal mutexes state so
not to stall irssi or hitting an assert() if no thread model is set for
the lib.

Hopefully, it would be nice to find a way to print text inside irssi
windows without the use of libperl or fixing the segfault when inside a
newly spawned thread.

Signed-off-by: David Goulet <dgoulet@ev0ke.net>